+++ src/main/java/com/seventhnode/youtubecrawler/service/YoutubeChannelService.java
@@ -1,16 +1,42 @@
 package com.seventhnode.youtubecrawler.service;
 
 
-import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.seventhnode.youtubecrawler.entity.YoutubeChannelInfo;
+import com.seventhnode.youtubecrawler.repository.YoutubeChannelRepository;
 
-public interface YoutubeChannelService {
+import java.util.List;
 
-    void save(YoutubeChannelInfo channelInfo);
-    void update(YoutubeChannelInfo channelInfo);
-    YoutubeChannelInfo get(long id);
-    YoutubeChannelInfo getByChannelId(String channelId);
-    List<YoutubeChannelInfo> getAll();
+@Service
+public class YoutubeChannelService {
 
+    @Autowired
+    YoutubeChannelRepository youtubeChannelRepository;
+
+
+    public void save(YoutubeChannelInfo channelInfo) {
+        youtubeChannelRepository.save(channelInfo);
+    }
+
+
+    public void update(YoutubeChannelInfo channelInfo) {
+        youtubeChannelRepository.save(channelInfo);
+    }
+
+
+    public YoutubeChannelInfo get(long id) {
+        return youtubeChannelRepository.findById(id).get();
+    }
+
+
+    public YoutubeChannelInfo getByChannelId(String channelId) {
+        return youtubeChannelRepository.findByChannelId(channelId);
+    }
+
+
+    public List<YoutubeChannelInfo> getAll() {
+        return youtubeChannelRepository.findAll();
+    }
 }
